requirements.txt
torch==2.2.1 
PyPDF2
python-docx
python-pptx
easyocr
transformers
librosa
pydub
sentencepiece
streamlit
moviepy
langchain
langchain_community
langchain_core
langchain_cohere
cohere
sentence-transformers
flask
langchain-chroma
